sandbox/gemini_practice/pyproject.toml
[project]
name = "gemini-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "jupyter>=1.1.1",
    "pillow>=11.2.1",
]
